# PERF: Series.fillna for pyarrow-backed dtypes 