--- a/pandas/core/arrays/arrow/array.py
+++ b/pandas/core/arrays/arrow/array.py
@@ -535,29 +535,38 @@ class ArrowExtensionArray(OpsMixin, ExtensionArray):
 
         value, method = validate_fillna_kwargs(value, method)
 
+        if is_array_like(value):
+            value = cast(ArrayLike, value)
+            if len(value) != len(self):
+                raise ValueError(
+                    f"Length of 'value' does not match. Got ({len(value)}) "
+                    f" expected {len(self)}"
+                )
+
+        def convert_fill_value(value, pa_type, dtype):
+            if value is None:
+                return value
+            if isinstance(value, (pa.Scalar, pa.Array, pa.ChunkedArray)):
+                return value
+            if is_array_like(value):
+                func = pa.array
+            else:
+                func = pa.scalar
+            try:
+                value = func(value, type=pa_type, from_pandas=True)
+            except pa.ArrowTypeError:
+                raise TypeError(f"Invalid value '{str(value)}' for dtype {dtype}")
+            return value
+
+        fill_value = convert_fill_value(value, self._data.type, self.dtype)
+
         try:
             if method is None and limit is None:
-                if is_array_like(value):
-                    val = cast(ArrayLike, value)
-                    if len(val) != len(self):
-                        raise ValueError(
-                            f"Length of 'value' does not match. Got ({len(val)}) "
-                            f" expected {len(self)}"
-                        )
-                    if not isinstance(val, (pa.Array, pa.ChunkedArray)):
-                        val = pa.array(val, type=self._data.type, from_pandas=True)
-                else:
-                    val = pa.scalar(value, type=self._data.type, from_pandas=True)
-                return type(self)(pc.fill_null(self._data, fill_value=val))
-
+                return type(self)(pc.fill_null(self._data, fill_value=fill_value))
             elif method == "pad" and limit is None and not pa_version_under7p0:
                 return type(self)(pc.fill_null_forward(self._data))
-
             elif method == "backfill" and limit is None and not pa_version_under7p0:
                 return type(self)(pc.fill_null_backward(self._data))
-
-        except (pa.ArrowInvalid, pa.ArrowTypeError):
-            raise TypeError(f"Invalid value '{str(value)}' for dtype {self.dtype}")
         except pa.ArrowNotImplementedError:
             # ArrowNotImplementedError: Function 'coalesce' has no kernel
             #   matching input types (duration[ns], duration[ns])