@mroeschke 2022-11-18
A few other places where this is showing up otherwise LGTM 

```
pandas/tests/extension/test_string.py::TestMissing::test_fillna_no_op_returns_copy[True-pyarrow]
pandas/tests/extension/test_string.py::TestMissing::test_fillna_no_op_returns_copy[False-pyarrow]
  /home/runner/work/pandas/pandas/pandas/tests/extension/base/missing.py:93: PerformanceWarning: Falling back on a non-pyarrow code path which may decrease performance. Upgrade to pyarrow >=7 to possibly suppress this warning.
    result = data.fillna(method="backfill")

pandas/tests/extension/test_string.py::TestMissing::test_fillna_series_method[ffill-True-pyarrow]
pandas/tests/extension/test_string.py::TestMissing::test_fillna_series_method[ffill-False-pyarrow]
pandas/tests/extension/test_string.py::TestMissing::test_fillna_series_method[bfill-True-pyarrow]
pandas/tests/extension/test_string.py::TestMissing::test_fillna_series_method[bfill-False-pyarrow]
  /home/runner/work/pandas/pandas/pandas/tests/extension/base/missing.py:123: PerformanceWarning: Falling back on a non-pyarrow code path which may decrease performance. Upgrade to pyarrow >=7 to possibly suppress this warning.
    result = pd.Series(data_missing).fillna(method=fillna_method)
```