@lukemanley 2022-11-19
> A few other places where this is showing up otherwise LGTM

fixed, thx