@lukemanley 2022-12-05
cc @mroeschke, @jbrockmendel 

It turns out `Block.fillna` does not dispatch to the underlying arrays in all cases including the simple case where method and limit are None. This highlights the impact of not dispatching for Arrow-backed arrays:

```
import pandas as pd
import pandas._testing as tm

N = 10**6
arr = pd.array(tm.rands_array(10, N), dtype="string[pyarrow]")
arr[::2] = None
ser = pd.Series(arr)

%timeit ser.fillna("foo")
60.5 ms ± 119 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)     # <- no dispatch

%timeit arr.fillna("foo")
4.89 ms ± 22.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)   # <- potential, if dispatched
``` 

It looks like the pattern might be to define `ArrowBackedExtensionBlock(ExtensionBlock)` with a `fillna` method that dispatches to `ArrowExtensionArray.fillna`.  However, the `Block.fillna` method is marked final. Any suggestions? Would it be problematic to remove the final attribute?