@mroeschke 2022-12-05
I imagine `ExtensionBlock` should generally define and just dispatch to to the array's `fillna` since (at least for the foreseeable future) they will be 1D?