@jbrockmendel 2022-12-05
> I imagine ExtensionBlock should generally define and just dispatch to to the array's fillna

That makes sense to me.  Would want to make sure we're not hurting perf in non-arrow cases.